+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+# 0.4.1
+
+- Increased rate for Navigraph AIRAC updates
+- Added transparency settings for OSM and satellite layers
+- Added transparency settings for weather layers, separated by light/dark theme
+- Added CartoDB with labels map layer
+- Improved airport dashboard loading performance
+- Added SimAware suffix support
+- Actions in overlays are now pinned when scrolling
+- Changed aircraft lines colors a bit for them to be more different from sector/tracon color
+
 # 0.4.0
 
 ## Highlights